back-requirements.txt
discord.py==2.2.2
bs4==0.0.1
datetime==4.3
tabulate==0.8.9
Pillow==10.3.0
pandas==1.4.0
lxml==4.9.3
python-fotmob-wrapper==0.1.8
requests==2.31.0
beautifulsoup4==4.10.0
pytest==8.1.1
pytest-asyncio==0.23.6
